BB-54 refactored name of storagerestapi package named 'repoistory' to 'repository'
