+++ src/main/java/com/butbetter/applicationservices/storagerestapi/service/StorageApiService.java
@@ -1,7 +1,7 @@
 package com.butbetter.applicationservices.storagerestapi.service;
 
 import com.butbetter.applicationservices.storagerestapi.model.ProductInformation;
-import com.butbetter.applicationservices.storagerestapi.repoistory.StorageOperation;
+import com.butbetter.applicationservices.storagerestapi.repository.StorageOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpEntity;
